+++ public/data/commentData.json
@@ -1,0 +1,17 @@
+[
+  {
+    "id": 1,
+    "username": "유저네임",
+    "userCommnet": "유저코멘트"
+  },
+  {
+    "id": 2,
+    "username": "유저네임",
+    "userCommnet": "유저코멘트"
+  },
+  {
+    "id": 3,
+    "username": "유저네임",
+    "userCommnet": "유저코멘트"
+  }
+]
