+++ public/data/feedData.json
@@ -1,0 +1,20 @@
+[
+  {
+    "userId": 1,
+    "profileName": "프로필이름1",
+    "profileUrl": "/image/jungsooKim/사.avif",
+    "feedContent": "피드컨텐츠1"
+  },
+  {
+    "userId": 2,
+    "profileName": "프로필이름2",
+    "profileUrl": "/image/jungsooKim/사.avif",
+    "feedContent": "피드컨텐츠2"
+  },
+  {
+    "userId": 3,
+    "profileName": "프로필이름3",
+    "profileUrl": "/image/jungsooKim/사.avif",
+    "feedContent": "피드컨텐츠3"
+  }
+]
