+++ src/pages/jungsooKim/Main/MainJung.scss
@@ -1,24 +1,8 @@
-.Login {
-}
 .changeFont2 {
   padding-left: 10px;
   border-left: 1px solid black;
   font-family: 'Lobster', cursive;
   font: size 25px;
-}
-
-.navLeft {
-  display: flex;
-  position: absolute;
-  left: 20%;
-}
-.navRight {
-  position: absolute;
-  right: 20%;
-}
-
-.SearchIcon {
-  left: 10px;
 }
 
 .icon {
@@ -28,21 +12,6 @@
 }
 input[placeholder] {
   text-align: center;
-}
-
-.SearchWrapper {
-  position: relative;
-
-  .fa-solid {
-    position: absolute;
-    left: 10px;
-    top: 5px;
-  }
-}
-
-body {
-  width: 100%;
-  height: 100%;
 }
 
 .moreIcon {
@@ -70,6 +39,8 @@ body {
 }
 
 .mainRight {
+  position: fixed;
+  right: 1%;
   float: right;
   width: 25%;
   height: 599px;
@@ -151,7 +122,7 @@ body {
   border-radius: 20px;
 }
 
-.discribe {
+.describe {
   align-self: center;
 }
 
@@ -159,14 +130,6 @@ body {
   position: absolute;
   right: 10px;
   color: #73b6f4;
-}
-
-.h1 {
-  margin: 0;
-}
-
-.h2 {
-  margin: 0;
 }
 
 .hour {
@@ -183,6 +146,7 @@ body {
   margin-top: 20px;
 }
 
-ul {
+li {
   list-style: none;
+  margin-left: 10px;
 }
